requirements.txt
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu117
py-enigma
tqdm
